backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
sentence-transformers
torch
faiss-cpu